Add gcc prerequisites to the build
Historically the build assumed that you could get gmp, mpc, and mpfr from
pkgsrc. For now we want to be still attempt to be able to build on systems which
can actually build the native illumos gate. Those systems however don't have
useful versions of the necessary libraries. Therefore, we add them to extra.

This does not add gcc yet for various reasons. It's worth mentioning for the
future that this builds, but not much more than that. It didn't get very far
into gcc because of target mismatch and well, needing a linker. You'll have to
set LD_LIBRARY_PATH to these suckers for gcc anyways.