Add support for building a basic arm-*-solaris2.11 gcc Change binutils target to be arm-*-solaris2.11
This folds in Rich Lowe's work to get a basic gcc up and running. Note that this
gcc is in actuality no different from an arm-*-none. However, for right now the
solaris specifics aren't any where near as important as the arm specifics and we
will slowly improve the compiler to be more solarisy.

Note that we ended up going with arm-*-solaris2.11 as a host tuple for the time
being. This is not meant to be a final be all end all decision, but rather an
expediency. One nice thing though, is that we can now build a basic gcc that,
using newlib, can actually be used to build just the kernel.