add a K&R-like cpp
This imports cpp from illumos-extra and makes an important change. Previously we
determined what preprocessor values it should emit based on the ones that are
emitted by levaraging what the compiler has bulit in. This doesn't work properly
for a cross arch build, so instead we specific TARGET_$(ARCH) to be the key for
what to emit as opposed to what the compiler's cpp would emit such as __i386__.