For now, don't trust the user's path.
Ideally this would come in via some other way probably whatever actually is
taking care of building the strap and then eventually illumos and then the
post-rendition of the strap. But until such a time we'll do this. It's ugly and
Keith would probably despair at it, but ah well.