add setrpath
This allows us to build the bootstrap compiler and not think too hard about how
we're going to jam the proper runpath of the destination in there or even what
it should be. As a result we end up using this to recurisvely set it for the
rest of the world that we care about.