gcc4: Default to the aapcs-linux ABI
For some reason, the changes in config/arm/sol2.h weren't enough for this.

The major (for us) difference between the aapcs and aapcs-linux ABIs is that
with strict AAPCS enums are 'short' rather than 'int', this means that a
function 'void foo(some_enum)' is not compatible with the prototype 'void
foo()', because 'some_enum' is a promotable type.  This, as you can imagine,
causes utter chaos.